Refactor. extract CLI prerequisites and arg validation into modules
